tests: consolidate config helpers, merge mixed-pool suite, lift imports
conftest had openai_config and azure_config, both returning OpenAIConfig
and differing only in whether base_url was set. They are not redundant --
the mock has separate respx routes for Azure's /openai/v1/... and direct
OpenAI's /v1/... -- but two similarly-named factories hid that. Merged
into openai_config(name, *, azure=True), so the distinction is visible at
the call site instead of encoded in which helper you picked.

test_mixed_pool.py folded into test_switchboard.py: both exercise
Switchboard, and the split was by fixture rather than by subject. Its
classes gain a TestMixedPool prefix, since names like TestSelection and
TestDispatch were unambiguous in their own file but not beside
TestSwitchboard's own selection tests.

All function-local imports lifted to module scope -- two of mine in
test_anthropic_api.py, three in the merged mixed-pool tests, and two
pre-existing `from .conftest import openai_config` inside test bodies in
test_switchboard.py. The shared Weather model is now defined once at
module level rather than redeclared in three test functions.

No behavior change: 87 tests, 100% coverage, pyright clean.
